+++ lib/mesh0.ml
@@ -186,16 +186,24 @@ let of_rows
       let faces = loop [] 0 0
       and bottom_cap () =
         let bot = List.hd layers in
-        let proj = Plane.project (Path3.to_plane ~no_check:true bot) in
-        let a = Util.array_of_list_map proj bot in
-        Triangulate.triangulate ~rev:(not rev) a
+        let hd = List.hd bot in
+        if List.for_all (V3.equal hd) bot
+        then []
+        else (
+          let proj = Plane.project (Path3.to_plane ~no_check:true bot) in
+          let a = Util.array_of_list_map proj bot in
+          Triangulate.triangulate ~rev:(not rev) a )
       and top_cap () =
         let offset = n_facets * (n_layers - 1)
         and top = List.nth layers (n_layers - 1) in
-        let proj = Plane.project (Path3.to_plane ~no_check:true top) in
-        let a = Util.array_of_list_map proj top in
-        let f = if rev then Util.offset_tri_rev else Util.offset_tri in
-        List.map (f offset) (Triangulate.triangulate a)
+        let hd = List.hd top in
+        if List.for_all (V3.equal hd) top
+        then []
+        else (
+          let proj = Plane.project (Path3.to_plane ~no_check:true top) in
+          let a = Util.array_of_list_map proj top in
+          let f = if rev then Util.offset_tri_rev else Util.offset_tri in
+          List.map (f offset) (Triangulate.triangulate a) )
       in
       match endcaps with
       | `Both -> List.concat [ top_cap (); bottom_cap (); faces ]
@@ -544,12 +552,12 @@ let mirror ax t = rev_faces { t with points = Array.map (V3.mirror ax) t.points 
 
 let to_binstl ~rev path { points = pts; faces } =
   let write_unsigned_long oc n =
-    Out_channel.output_byte oc ((n lsr 0) land 0xFF);
-    Out_channel.output_byte oc ((n lsr 8) land 0xFF);
-    Out_channel.output_byte oc ((n lsr 16) land 0xFF);
-    Out_channel.output_byte oc ((n lsr 24) land 0xFF)
+    Out_channel.output_byte oc Int32.(to_int n lsr 0);
+    Out_channel.output_byte oc Int32.(to_int n lsr 8);
+    Out_channel.output_byte oc Int32.(to_int n lsr 16);
+    Out_channel.output_byte oc Int32.(to_int @@ shift_right_logical n 24)
   in
-  let write_float oc n = write_unsigned_long oc Int32.(to_int @@ bits_of_float n) in
+  let write_float oc n = write_unsigned_long oc (Int32.bits_of_float n) in
   let write_v3 oc p =
     write_float oc @@ V3.x p;
     write_float oc @@ V3.y p;
@@ -563,7 +571,7 @@ let to_binstl ~rev path { points = pts; faces } =
     Bytes.set bs 79 '\n';
     Out_channel.output_bytes oc bs;
     (* number of facets *)
-    write_unsigned_long oc (List.length faces);
+    write_unsigned_long oc (Int32.of_int @@ List.length faces);
     List.iter
       (fun (a, b, c) ->
         let a, b, c =
@@ -615,6 +623,106 @@ let to_stl ~rev path { points = pts; faces } =
   in
   Out_channel.with_open_bin path f
 
-let to_stl ?(ascii = false) ?(rev = true) ?eps path t =
-  let t = merge_points ?eps t in
+let to_stl ?(ascii = false) ?(rev = true) path t =
   if ascii then to_stl ~rev path t else to_binstl ~rev path t
+
+let of_binstl_ic ?(rev = true) ?eps ic =
+  let read_unsigned_long ic =
+    let ch1 = input_byte ic
+    and ch2 = input_byte ic
+    and ch3 = input_byte ic
+    and big = Int32.shift_left (Int32.of_int (input_byte ic)) 24 in
+    let base = Int32.of_int (ch1 lor (ch2 lsl 8) lor (ch3 lsl 16)) in
+    Int32.logor base big
+  in
+  let read_float ic = Int32.float_of_bits (read_unsigned_long ic) in
+  let read_vertex ic =
+    let x = read_float ic
+    and y = read_float ic
+    and z = read_float ic in
+    V3.v x y z
+  in
+  let rec loop ic n_facets i pos ps tris =
+    if i < n_facets
+    then (
+      (* skip over normal *)
+      In_channel.seek ic (Int64.of_int (pos + 12));
+      let a = read_vertex ic
+      and b = read_vertex ic
+      and c = read_vertex ic in
+      (* normal + verts = 48, skip over 2 \000 property bytes to 50 *)
+      In_channel.seek ic (Int64.of_int (pos + 50));
+      let j = i * 3 in
+      let tri = if rev then j + 2, j + 1, j else j, j + 1, j + 2 in
+      loop ic n_facets (i + 1) (pos + 50) (c :: b :: a :: ps) (tri :: tris) )
+    else ps, tris
+  in
+  In_channel.seek ic (Int64.of_int 80);
+  let n_facets = Int32.to_int @@ read_unsigned_long ic in
+  if Int64.(In_channel.length ic <> of_int ((n_facets * 50) + 84))
+  then failwith "Invalid binary stl (file size <> facets * 50 + 84)";
+  (* header 80 bytes + 4 bytes for n_facets -> position 84 *)
+  let ps, faces = loop ic n_facets 0 84 [] [] in
+  merge_points ?eps { points = Util.array_of_list_rev ps; faces }
+
+let of_asciistl_ic ?(rev = true) ?eps ic =
+  let validate_line ic prefix =
+    match In_channel.input_line ic with
+    | Some line ->
+      if not @@ String.(starts_with ~prefix @@ trim line)
+      then failwith (Printf.sprintf "Invalid block tag in ascii stl (expected %s)" prefix)
+    | None ->
+      failwith (Printf.sprintf "Unexpected end of ascii stl (expected %s line)" prefix)
+  in
+  let read_vertex ic =
+    match In_channel.input_line ic with
+    | Some line ->
+      ( try
+          let[@warning "-partial-match"] [ "vertex"; x; y; z ] =
+            String.(split_on_char ' ' @@ trim line)
+          in
+          Float.(V3.v (of_string x) (of_string y) (of_string z))
+        with
+        | _ -> failwith "Invalid vertex encountered in ascii stl." )
+    | None -> failwith "Unexpected end of ascii stl file."
+  in
+  let rec loop ic i ps tris =
+    match In_channel.input_line ic with
+    | Some line ->
+      let line = String.trim line in
+      if String.starts_with ~prefix:"endsolid" line
+      then ps, tris
+      else if String.starts_with ~prefix:"facet" line
+      then (
+        (* ignoring normal which comes after facet *)
+        validate_line ic "outer loop";
+        let a = read_vertex ic
+        and b = read_vertex ic
+        and c = read_vertex ic in
+        let tri = if rev then i + 2, i + 1, i else i, i + 1, i + 2 in
+        validate_line ic "endloop";
+        validate_line ic "endfacet";
+        loop ic (i + 3) (c :: b :: a :: ps) (tri :: tris) )
+      else if String.length line = 0
+      then loop ic i ps tris
+      else failwith "Invalid block tag in ascii stl (expected facet)."
+    | None -> failwith "Unexpected end of ascii stl (expected facet or end of solid)"
+  in
+  if In_channel.length ic < Int64.of_int 15
+  then failwith "File too short to be a valid ascii stl (< 15 bytes)";
+  In_channel.seek ic (Int64.of_int 0);
+  validate_line ic "solid";
+  let ps, faces = loop ic 0 [] [] in
+  merge_points ?eps { points = Util.array_of_list_rev ps; faces }
+
+let of_stl ?rev ?eps path =
+  let is_ascii ic =
+    match In_channel.input_line ic with
+    | Some line when String.(starts_with ~prefix:"solid " @@ trim line) ->
+      ( match In_channel.input_line ic with
+        | Some line -> String.(starts_with ~prefix:"facet " @@ trim line)
+        | _ -> false )
+    | _ -> false
+  in
+  In_channel.with_open_bin path (fun ic ->
+    if is_ascii ic then of_asciistl_ic ?rev ?eps ic else of_binstl_ic ?rev ?eps ic )